renamed bottombarviewmodel to bottomnavigationbarviewmodel and moved tab to own file
