+++ app/src/main/kotlin/com/platform/openemoji/navigation/BottomNavigationBarTab.kt
@@ -1,0 +1,6 @@
+package com.platform.openemoji.navigation
+
+data class BottomNavigationBarTab(
+    val orderValue: Int,
+    val route: String,
+)
